requirements.txt
tqdm==4.33.0
urllib3==1.26.5
requests==2.25.1
python-dotenv==0.19.2
Flask==3.0.3